11.12 결제 완성

+++ src/main/java/com/kosateam2/chicken/dto/Order.java
@@ -39,13 +39,13 @@ public class Order {
 	public double getLat() {
 		return lat;
 	}
-	public void setLat(int lat) {
+	public void setLat(double lat) {
 		this.lat = lat;
 	}
 	public double getLng() {
 		return lng;
 	}
-	public void setLng(int lng) {
+	public void setLng(double lng) {
 		this.lng = lng;
 	}
 	public Date getDate() {
